refactor(core): simplify module paths and unify file naming 2
