+++ novel_downloader/core/downloaders/__init__.py
@@ -11,8 +11,7 @@ Each downloader is responsible for orchestrating the full lifecycle
 of retrieving, parsing, and saving novel content for a given source.
 """
 
-from .common import CommonDownloader
-from .common_async import CommonAsyncDownloader
+from .common import CommonAsyncDownloader, CommonDownloader
 from .qidian import QidianDownloader
 
 __all__ = [
