+++ novel_downloader/core/downloaders/base/__init__.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+novel_downloader.core.downloaders.base
+--------------------------------------
+
+"""
+
+from .base_async import BaseAsyncDownloader
+from .base_sync import BaseDownloader
+
+__all__ = [
+    "BaseAsyncDownloader",
+    "BaseDownloader",
+]
