+++ novel_downloader/core/downloaders/common/__init__.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+novel_downloader.core.downloaders.common
+----------------------------------------
+
+"""
+
+from .common_async import CommonAsyncDownloader
+from .common_sync import CommonDownloader
+
+__all__ = [
+    "CommonAsyncDownloader",
+    "CommonDownloader",
+]
