+++ novel_downloader/core/downloaders/qidian/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+novel_downloader.core.downloaders.qidian
+----------------------------------------
+
+"""
+
+from .qidian_sync import QidianDownloader
+
+__all__ = ["QidianDownloader"]
