+++ novel_downloader/core/parsers/common/helper.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-novel_downloader.core.parsers.common_parser.helpers
----------------------------------------------------
+novel_downloader.core.parsers.common.helpers
+--------------------------------------------
 
 Shared utility functions for parsing Common pages.
 """
