+++ novel_downloader/core/requesters/common/__init__.py
@@ -1,16 +1,16 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-novel_downloader.core.requesters.common_requester
--------------------------------------------------
+novel_downloader.core.requesters.common
+---------------------------------------
 
 This module provides the `CommonSession` class wrapper for common HTTP
 request operations to novel websites. It serves as a unified access
 point to import `CommonSession` without exposing lower-level modules.
 """
 
-from .common_async_session import CommonAsyncSession
-from .common_session import CommonSession
+from .async_session import CommonAsyncSession
+from .session import CommonSession
 
 __all__ = [
     "CommonAsyncSession",
