+++ novel_downloader/core/savers/common/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-novel_downloader.core.savers.common_saver
+novel_downloader.core.savers.common
 -----------------------------------------
 
 This module provides the `CommonSaver` class for handling the saving process
